+++ Competition_Code/src/lib2131/pid.cpp
@@ -3,12 +3,12 @@
 namespace lib2131
 {
 PID::PID() : kP(0), kI(0), kD(0), P(0), I(0), D(0), lastP(0), lastI(0), lastD(0) {}
-PID::PID(double KP, double KI, double KD)
+PID::PID(float KP, float KI, float KD)
     : kP(KP), kI(KI), kD(KD), P(0), I(0), D(0), lastP(0), lastI(0), lastD(0)
 {
 }
 
-double PID::calc(double error, int dTime)
+float PID::calc(float error, int dTime)
 {
   P = error;
   I += P * dTime;
